Update all route
Vì ko thể check role ở session nên là ở mỗi route đều có đoạn check riêng

+++ Code_copy_ver06/backend/routes/event.js
@@ -4,6 +4,7 @@ var router = express.Router();
 //import model before use
 var EventModel = require('../models/EventModel');
 const FacultyModel = require('../models/FacultyModel');
+var UserModel = require('../models/UserModel');
 const {checkAdminSession, verifyToken} = require('../middlewares/auth');
 //------------------------------------------------------------------------
 //show all 
@@ -14,6 +15,7 @@ router.get('/', verifyToken, async(req, res) => {
         if(!userData){
             return res.status(400).json({success: false, error: "Not found user"});
         }
+        
         const userRole = userData.role.toString();
         if(userRole === '65e61d9bb8171b6e90f92da3'){
             //Code ở đây--------------------------
@@ -33,15 +35,27 @@ router.get('/', verifyToken, async(req, res) => {
 
 //-----------------------------------------------------------------------
 //delete specific 
-router.delete('/delete/:id', verifyToken, checkAdminSession, async (req, res) => {
+router.delete('/delete/:id', verifyToken, async (req, res) => {
     try {
-        const eventId = req.params.id;
-        const deletedEvent = await EventModel.findByIdAndDelete(eventId);
-        if (!deletedEvent) {
-            res.status(404).json({ success: false, error: "Event not found" });
-            return;
+        const userId = req.userId;
+        const userData = await UserModel.findById(userId);
+        if(!userData){
+            return res.status(400).json({success: false, error: "Not found user"});
         }
-        res.status(200).json({ success: true, message: "Event deleted successfully" });
+        const userRole = userData.role.toString();
+        if(userRole === '65e61d9bb8171b6e90f92da3'){
+            //Code ở đây--------------------------
+            const eventId = req.params.id;
+            const deletedEvent = await EventModel.findByIdAndDelete(eventId);
+            if (!deletedEvent) {
+                res.status(404).json({ success: false, error: "Event not found" });
+                return;
+            }
+            res.status(200).json({ success: true, message: "Event deleted successfully" });
+            //----------------------------------
+        } else {
+            return res.status(400).json({ success: false, error: "Not right Role" });
+        }
     } catch (error) {
         console.error("Error while deleting event:", error);
         res.status(500).json({ success: false, error: "Internal Server Error" });
@@ -51,10 +65,23 @@ router.delete('/delete/:id', verifyToken, checkAdminSession, async (req, res) =>
 //------------------------------------------------------------------------
 //create 
 //render form for user to input
-router.get('/add', verifyToken, checkAdminSession, async (req, res) => {
+router.get('/add', verifyToken,  async (req, res) => {
     try{
-        var facultyList = await FacultyModel.find({});
-        res.status(200).json({ success: true, message: "Render add event form", data: facultyList });
+        const userId = req.userId;
+        const userData = await UserModel.findById(userId);
+        if(!userData){
+            return res.status(400).json({success: false, error: "Not found user"});
+        }
+        const userRole = userData.role.toString();
+        if(userRole === '65e61d9bb8171b6e90f92da3'){
+            //Code ở đây--------------------------
+            var facultyList = await FacultyModel.find({});
+            res.status(200).json({ success: true, message: "Render add event form", data: facultyList });
+            //----------------------------------
+        } else {
+            return res.status(400).json({ success: false, error: "Not right Role" });
+        }
+//-------------------------------------------------
     }catch(error){
         console.error("Error while making event:", error);
         res.status(500).send("Internal Server Error");
@@ -62,22 +89,35 @@ router.get('/add', verifyToken, checkAdminSession, async (req, res) => {
 });
 
 //receive form data and insert it to database
-router.post('/add', verifyToken, checkAdminSession, async (req, res) => {
+router.post('/add', verifyToken,  async (req, res) => {
     //get value by form : req.body
     try{
-        const requirement = req.body.requirement;
-        const deadline1 = req.body.deadline1;
-        const deadline2 = req.body.deadline2;
-        const faculty = req.body.faculty;
-        await EventModel.create(
-            {
-                requirement: requirement,
-                deadline1:deadline1,
-                deadline2:deadline2,
-                faculty: faculty,
-            }
-        );
-        res.status(201).json({ success: true, message: "Event created successfully" });
+        const userId = req.userId;
+        const userData = await UserModel.findById(userId);
+        if(!userData){
+            return res.status(400).json({success: false, error: "Not found user"});
+        }
+        const userRole = userData.role.toString();
+        if(userRole === '65e61d9bb8171b6e90f92da3'){
+            //Code ở đây--------------------------
+            const requirement = req.body.requirement;
+            const deadline1 = req.body.deadline1;
+            const deadline2 = req.body.deadline2;
+            const faculty = req.body.faculty;
+            await EventModel.create(
+                {
+                    requirement: requirement,
+                    deadline1:deadline1,
+                    deadline2:deadline2,
+                    faculty: faculty,
+                }
+            );
+            res.status(201).json({ success: true, message: "Event created successfully" });
+            //----------------------------------
+        } else {
+            return res.status(400).json({ success: false, error: "Not right Role" });
+        }
+//-------------------------------------------------
     } catch (error) {
         if (error.name === 'ValidationError') {
            let InputErrors = {};
@@ -92,37 +132,62 @@ router.post('/add', verifyToken, checkAdminSession, async (req, res) => {
 
 //---------------------------------------------------------------------------
 //edit 
-router.get('/edit/:id', verifyToken, checkAdminSession, async (req, res) => {
+router.get('/edit/:id', verifyToken,  async (req, res) => {
     try{
-        const id = req.params.id;
-        const event = await EventModel.findById(id).populate('faculty');
-        if (!event) {
-            res.status(400).send("Event not found");
+        const userId = req.userId;
+        const userData = await UserModel.findById(userId);
+        if(!userData){
+            return res.status(400).json({success: false, error: "Not found user"});
         }
-        const facultyList = await FacultyModel.find({});
-        res.status(200).json({ success: true, message: "Render edit event form", event, facultyList });
+        const userRole = userData.role.toString();
+        if(userRole === '65e61d9bb8171b6e90f92da3'){
+            //Code ở đây--------------------------
+            const id = req.params.id;
+            const event = await EventModel.findById(id).populate('faculty');
+            if (!event) {
+                res.status(400).send("Event not found");
+            }
+            const facultyList = await FacultyModel.find({});
+            res.status(200).json({ success: true, message: "Render edit event form", event, facultyList });
+            //----------------------------------
+        } else {
+            return res.status(400).json({ success: false, error: "Not right Role" });
+        }
+//-------------------------------------------------
     }catch(error){
         console.error("Error while editing event:", error);
         res.status(500).send("Internal Server Error");
     }
-    
 });
 
-router.put('/edit/:id', verifyToken, checkAdminSession, async(req, res) => {
+router.put('/edit/:id', verifyToken,  async(req, res) => {
     try{
-        const id = req.params.id;
-        const event = await EventModel.findById(id);
-        if (!event) {
-            res.status(400).send("Event not found");
+        const userId = req.userId;
+        const userData = await UserModel.findById(userId);
+        if(!userData){
+            return res.status(400).json({success: false, error: "Not found user"});
         }
-        // Update student details
-        event.requirement = req.body.requirement;
-        event.deadline1 = req.body.deadline1;
-        event.deadline2 = req.body.deadline2;
-        event.faculty = req.body.faculty;
-
-        await event.save();
-        res.status(200).json({ success: true, message: "Create Event successfully" });
+        const userRole = userData.role.toString();
+        if(userRole === '65e61d9bb8171b6e90f92da3'){
+            //Code ở đây--------------------------
+            const id = req.params.id;
+            const event = await EventModel.findById(id);
+            if (!event) {
+                res.status(400).send("Event not found");
+            }
+            // Update student details
+            event.requirement = req.body.requirement;
+            event.deadline1 = req.body.deadline1;
+            event.deadline2 = req.body.deadline2;
+            event.faculty = req.body.faculty;
+    
+            await event.save();
+            res.status(200).json({ success: true, message: "Create Event successfully" });
+            //----------------------------------
+        } else {
+            return res.status(400).json({ success: false, error: "Not right Role" });
+        }
+//-------------------------------------------------
     } catch (error) {
         if (error.name === 'ValidationError') {
            let InputErrors = {};
